Use loadtest4j 0.16.0 (#40)

+++ src/main/scala/org/loadtest4j/drivers/gatling/GatlingResponseTime.scala
@@ -5,7 +5,7 @@ import java.time.Duration
 import org.loadtest4j.driver.DriverResponseTime
 
 class GatlingResponseTime(percentile: Double => Int) extends DriverResponseTime {
-  override def getPercentile(i: Int): Duration = {
+  override def getPercentile(i: Double): Duration = {
     val millis = percentile(i)
     Duration.ofMillis(millis)
   }
